render_preview is opt-in
