+++ test/sandbox/test/components/without_preview_component_test.rb
@@ -3,6 +3,8 @@
 require "test_helper"
 
 class WithoutPreviewComponentTest < ViewComponent::TestCase
+  include ViewComponent::RenderPreviewHelper
+
   def test_render_preview
     error = assert_raises NameError do
       render_preview(:default)
